// src/app/shared/Models/Perfil.ts
export class Profile {

    id?: number;
    name?: string;
    attempts?: number;
    expirationDay?: number;
    cantHistory?: number;
    characterUpper?: Boolean;
    characterLower?: Boolean;
    characterNumber?: Boolean;
    characterSpecial?: Boolean;
    lengthMinimum?: number;
    idStatus?: number;
    boolDelete?: Boolean;
    startIdUser? : string;
    updateIdUser? : string;

}